improved output indformation and fixed bug
Added "?" to the output file to indicate when we were unable to test a clan in a tree. Up until now, this would have been a "1" as it did not violate the CLan, however this change may be more useful to users as (for instnace) it will allow to identify where trees were not tested by any Clan.
Also fixed a bug, where for a very small clade a Clan may be passed when in fact it violates it.